requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
google-adk>=1.2.0
google-cloud-firestore>=2.19.0
python-dotenv>=1.0.0
